refactor(shell): extract shell utilities and tests
Extracted shell-related utility functions (`splitCommands`, `getCommandRoot`, `getCommandRoots`, `stripShellWrapper`, `detectCommandSubstitution`, `isCommandAllowed`) from `packages/core/src/tools/shell.ts` into a new file `packages/core/src/utils/shell-utils.ts`.

This refactoring improves code organization and separation of concerns by isolating the pure utility functions from the `ShellTool` class implementation.

Additionally, the corresponding tests for these utility functions have been moved from `packages/core/src/tools/shell.test.ts` to a new test file `packages/core/src/utils/shell-utils.test.ts`. The original `shell.test.ts` now only contains tests relevant to the `ShellTool` class itself.
